skillswap-backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-cors-headers>=4.6,<4.7
channels[daphne]>=4.2,<4.3
psycopg[binary]>=3.2,<3.3
